chore: add requirements.txt
